requirements.txt
deepface~=0.0.98
opencv-python~=4.13.0.90
numpy~=2.3.5
tensorflow
retina-face
tf-keras
librosa~=0.11.0
sounddevice~=0.5.5
openai~=2.16.0
pillow~=12.1.0
openai-whisper
python-dotenv~=1.2.1